+++ code/IRRMeshLoader.cpp
@@ -422,7 +422,7 @@ void IRRMeshImporter::InternReadFile( const std::string& pFile,
 					}
 					if (!curIdx)	{
 						curFace->mNumIndices = 3;
-						curFace->mIndices = new unsigned int[3];
+						curFace->mIndices = new size_t[3];
 					}
 
 					unsigned int idx = strtoul10(sz,&sz);
