+++ code/MD3Loader.cpp
@@ -966,7 +966,7 @@ void MD3Importer::InternReadFile( const std::string& pFile,
 		// Fill in all triangles
 		unsigned int iCurrent = 0;
 		for (unsigned int i = 0; i < (unsigned int)pcSurfaces->NUM_TRIANGLES;++i)	{
-			pcMesh->mFaces[i].mIndices = new unsigned int[3];
+			pcMesh->mFaces[i].mIndices = new size_t[3];
 			pcMesh->mFaces[i].mNumIndices = 3;
 
 			//unsigned int iTemp = iCurrent;
