+++ code/MDLLoader.cpp
@@ -446,7 +446,7 @@ void MDLImporter::InternReadFile_Quake1( )
 	unsigned int iCurrent = 0;
 	for (unsigned int i = 0; i < (unsigned int) pcHeader->num_tris;++i)
 	{
-		pcMesh->mFaces[i].mIndices = new unsigned int[3];
+		pcMesh->mFaces[i].mIndices = new size_t[3];
 		pcMesh->mFaces[i].mNumIndices = 3;
 
 		unsigned int iTemp = iCurrent;
@@ -657,7 +657,7 @@ void MDLImporter::InternReadFile_3DGS_MDL345( )
 		// now iterate through all triangles
 		unsigned int iCurrent = 0;
 		for (unsigned int i = 0; i < (unsigned int) pcHeader->num_tris;++i)	{
-			pcMesh->mFaces[i].mIndices = new unsigned int[3];
+			pcMesh->mFaces[i].mIndices = new size_t[3];
 			pcMesh->mFaces[i].mNumIndices = 3;
 
 			unsigned int iTemp = iCurrent;
@@ -712,7 +712,7 @@ void MDLImporter::InternReadFile_3DGS_MDL345( )
 		// now iterate through all triangles
 		unsigned int iCurrent = 0;
 		for (unsigned int i = 0; i < (unsigned int) pcHeader->num_tris;++i)	{
-			pcMesh->mFaces[i].mIndices = new unsigned int[3];
+			pcMesh->mFaces[i].mIndices = new size_t[3];
 			pcMesh->mFaces[i].mNumIndices = 3;
 
 			unsigned int iTemp = iCurrent;
@@ -1822,9 +1822,9 @@ void MDLImporter::GenerateOutputMeshes_3DGS_MDL7(
 
 			// iterate through all faces and build an unique set of vertices
 			unsigned int iCurrent = 0;
-			for (unsigned int iFace = 0; iFace < pcMesh->mNumFaces;++iFace)	{
+			for (size_t iFace = 0; iFace < pcMesh->mNumFaces;++iFace)	{
 				pcMesh->mFaces[iFace].mNumIndices = 3;
-				pcMesh->mFaces[iFace].mIndices = new unsigned int[3];
+				pcMesh->mFaces[iFace].mIndices = new size_t[3];
 
 				unsigned int iSrcFace = splitGroupData.aiSplit[i]->operator[](iFace);
 				const MDL::IntFace_MDL7& oldFace = groupData.pcFaces[iSrcFace];
