+++ code/ProcessHelper.cpp
@@ -323,7 +323,7 @@ aiMesh* MakeSubmesh(const aiMesh *pMesh, const std::vector<unsigned int> &subMes
 		const aiFace& srcFace = pMesh->mFaces[subMeshFaces[a]];
 		aiFace& dstFace = oMesh->mFaces[a];
 		dstFace.mNumIndices = srcFace.mNumIndices;
-		dstFace.mIndices = new unsigned int[dstFace.mNumIndices];
+		dstFace.mIndices = new size_t[dstFace.mNumIndices];
 
 		// accumulate linearly all the vertices of the source face
 		for( size_t b = 0; b < dstFace.mNumIndices; ++b )	{
